Bump blockifier to support Cairo 2.10.0 (#2866)
<!-- Reference any GitHub issues resolved by this PR -->

Closes #2862

## Introduced changes

<!-- A brief description of the changes -->

-

## Checklist

<!-- Make sure all of these are complete -->

- [ ] Linked relevant issue
- [ ] Updated relevant documentation
- [ ] Added relevant tests
- [ ] Performed self-review of the code
- [ ] Added changes to `CHANGELOG.md`

---------

+++ crates/cheatnet/src/runtime_extensions/call_to_blockifier_runtime_extension/mod.rs
@@ -1,6 +1,5 @@
 use std::marker::PhantomData;
 
-use crate::constants::TEST_ADDRESS;
 use blockifier::execution::entry_point::{CallEntryPoint, CallType};
 use blockifier::execution::execution_utils::felt_from_ptr;
 use blockifier::execution::syscalls::{
@@ -17,8 +16,8 @@ use blockifier::execution::{
 use cairo_vm::types::relocatable::MaybeRelocatable;
 use cairo_vm::vm::{errors::hint_errors::HintError, vm_core::VirtualMachine};
 use runtime::{ExtendedRuntime, ExtensionLogic, SyscallHandlingResult, SyscallPtrAccess};
+use starknet_api::contract_class::EntryPointType;
 use starknet_api::core::ContractAddress;
-use starknet_api::deprecated_contract_class::EntryPointType;
 
 use crate::state::CheatnetState;
 
@@ -32,6 +31,7 @@ use crate::runtime_extensions::call_to_blockifier_runtime_extension::{
 
 use super::cheatable_starknet_runtime_extension::CheatableStarknetRuntime;
 use conversions::string::TryFromHexStr;
+use runtime::starknet::constants::TEST_ADDRESS;
 
 pub mod execution;
 pub mod panic_data;
@@ -111,7 +111,7 @@ impl ExecuteCall for CallContractRequest {
             storage_address: contract_address,
             caller_address: TryFromHexStr::try_from_hex_str(TEST_ADDRESS).unwrap(),
             call_type: CallType::Call,
-            initial_gas: u64::MAX,
+            initial_gas: i64::MAX as u64,
         };
 
         call_entry_point(
